Add neoconf integration
